+++ crates/biome_service/src/file_handlers/json.rs
@@ -65,7 +65,7 @@ impl Language for JsonLanguage {
         let indent_width = if let Some(indent_width) = language.indent_width {
             indent_width
         } else {
-            global.indent_size.unwrap_or_default()
+            global.indent_width.unwrap_or_default()
         };
         JsonFormatOptions::default()
             .with_indent_style(indent_style)